Snapshot 202403

+++ src/metadata/identify_tools.cpp
@@ -1,5 +1,5 @@
 /* -*- C++ -*-
- * Copyright 2019-2021 LibRaw LLC (info@libraw.org)
+ * Copyright 2019-2024 LibRaw LLC (info@libraw.org)
  *
  LibRaw uses code from dcraw.c -- Dave Coffin's raw photo decoder,
  dcraw.c is copyright 1997-2018 by Dave Coffin, dcoffin a cybercom o net.
@@ -94,7 +94,7 @@ void LibRaw::remove_trailing_spaces(char *string, size_t len)
   if (len < 3)
     return; // also not needed
   len = strnlen(string, len - 1);
-  for (int i = int(len) - 1; i >= 0; i--)
+  for (size_t i = len - 1; i >= 0; i--)
   {
     if (isspace((unsigned char)string[i]))
       string[i] = 0;
